base_product renamed to base_classes
